refactor(frontend): decompose agent studio shell model
