+++ packages/frontend/src/pages/agents/shell/use-agents-page-shell-model.tsx
@@ -1,22 +1,6 @@
-import type { AgentRole } from "@openducktor/core";
-import { memo, type ReactElement, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { useNavigationType, useSearchParams } from "react-router-dom";
+import { type ReactElement, useCallback, useRef, useState } from "react";
 import { SessionStartModal } from "@/components/features/agents";
-import { MemoizedAgentStudioRightPanel } from "@/components/features/agents/agent-studio-right-panel";
-import type {
-  ActiveTaskSessionContextByTaskId,
-  KanbanTaskSession,
-} from "@/components/features/kanban/kanban-task-activity";
-import { MergedPullRequestConfirmDialog } from "@/components/features/pull-requests/merged-pull-request-confirm-dialog";
-import {
-  TaskDetailsSheetController,
-  type TaskDetailsSheetControllerHandle,
-} from "@/components/features/task-details/task-details-sheet-controller";
-import type { BuildToolsSessionDescriptor } from "@/features/agent-studio-build-tools/use-agent-studio-build-tools-bootstrap";
-import { useAgentStudioBuildWorktreeRefresh } from "@/features/agent-studio-build-tools/use-agent-studio-build-worktree-refresh";
-import type { GitDiffRefresh } from "@/features/agent-studio-git";
 import { HumanReviewFeedbackModal } from "@/features/human-review-feedback/human-review-feedback-modal";
-import { toAgentSessionSummary } from "@/state/agent-sessions-store";
 import {
   useChecksOperationsContext,
   useRuntimeAvailabilityContext,
@@ -28,30 +12,13 @@ import {
   useTasksState,
   useWorkspaceState,
 } from "@/state/app-state-provider";
-import type { AgentStudioQueryUpdate } from "../agent-studio-navigation";
-import {
-  type AgentStudioOrchestrationSelectionContext,
-  useAgentStudioOrchestrationController,
-} from "../use-agent-studio-orchestration-controller";
-import { useAgentStudioQuerySessionSync } from "../use-agent-studio-query-session-sync";
-import { useAgentStudioQuerySync } from "../use-agent-studio-query-sync";
-import { useAgentStudioRebaseConflictResolution } from "../use-agent-studio-rebase-conflict-resolution";
-import { useAgentStudioSelectionController } from "../use-agent-studio-selection-controller";
-import { useAgentStudioReadiness } from "../use-agents-page-readiness";
-import {
-  type AgentStudioGitConflictQuickActionContext,
-  type UseAgentsPageRightPanelModelArgs,
-  useAgentsPageRightPanelModel,
-} from "../use-agents-page-right-panel-model";
-import {
-  type AgentStudioSelectionIntent,
-  isSelectionIntentResolved,
-} from "./agent-studio-selection-intent";
+import type { useAgentStudioOrchestrationController } from "../use-agent-studio-orchestration-controller";
+import type { AgentStudioGitConflictQuickActionContext } from "../use-agents-page-right-panel-model";
 import { gitConflictQuickActionContextsEqual } from "./git-conflict-quick-action-context";
-import {
-  useForwardedWorktreeRefresh,
-  type WorktreeRefreshRef,
-} from "./use-forwarded-worktree-refresh";
+import { useAgentsPageOrchestrationShellModel } from "./use-agents-page-orchestration-shell-model";
+import { useAgentsPageRightPanelShellModel } from "./use-agents-page-right-panel-shell-model";
+import { useAgentsPageRouteSessionModel } from "./use-agents-page-route-session-model";
+import { useAgentsPageShellOverlays } from "./use-agents-page-shell-overlays";
 
 type AgentsPageShellModel = {
   activeWorkspace: ReturnType<typeof useWorkspaceState>["activeWorkspace"];
@@ -80,62 +47,6 @@ type AgentsPageShellModel = {
   taskDetailsSheet: ReactElement;
 };
 
-const AgentsPageRightPanelRuntime = memo(function AgentsPageRightPanelRuntime({
-  refreshWorktreeRef,
-  ...args
-}: UseAgentsPageRightPanelModelArgs & {
-  refreshWorktreeRef: WorktreeRefreshRef;
-}): ReactElement | null {
-  const { rightPanelModel, refreshWorktree } = useAgentsPageRightPanelModel(args);
-
-  useEffect(() => {
-    refreshWorktreeRef.current = refreshWorktree;
-    return () => {
-      if (refreshWorktreeRef.current === refreshWorktree) {
-        refreshWorktreeRef.current = null;
-      }
-    };
-  }, [refreshWorktree, refreshWorktreeRef]);
-
-  return rightPanelModel ? <MemoizedAgentStudioRightPanel model={rightPanelModel} /> : null;
-});
-
-function AgentsPageBuildWorktreeRefreshRuntime({
-  panelKind,
-  isPanelOpen,
-  viewRole,
-  activeSession,
-  isSessionHistoryHydrating,
-  refreshWorktreeRef,
-}: {
-  panelKind: "documents" | "build_tools" | null;
-  isPanelOpen: boolean;
-  viewRole: UseAgentsPageRightPanelModelArgs["viewRole"];
-  activeSession: AgentStudioOrchestrationSelectionContext["viewActiveSession"];
-  isSessionHistoryHydrating: boolean;
-  refreshWorktreeRef: WorktreeRefreshRef;
-}): null {
-  const refreshWorktree = useForwardedWorktreeRefresh(refreshWorktreeRef);
-
-  useAgentStudioBuildWorktreeRefresh({
-    viewRole: panelKind === "build_tools" && isPanelOpen ? viewRole : null,
-    activeSession,
-    isSessionHistoryHydrating,
-    refreshWorktree,
-  });
-
-  return null;
-}
-
-const EMPTY_TASK_SESSIONS_BY_TASK_ID = new Map<string, KanbanTaskSession[]>();
-const EMPTY_ACTIVE_TASK_SESSION_CONTEXT_BY_TASK_ID: ActiveTaskSessionContextByTaskId = new Map();
-
-const noopOpenSession = (
-  _taskId: string,
-  _role: AgentRole,
-  _options?: { externalSessionId?: string | null },
-): void => {};
-
 export function useAgentsPageShellModel(): AgentsPageShellModel {
   const { activeBranch, branches, activeWorkspace } = useWorkspaceState();
   const workspaceRepoPath = activeWorkspace?.repoPath ?? null;
@@ -161,70 +72,46 @@ export function useAgentsPageShellModel(): AgentsPageShellModel {
     unlinkingPullRequestTaskId,
     setTaskTargetBranch,
   } = useTasksState();
+  const agentOperations = useAgentOperations();
   const {
-    bootstrapTaskSessions,
     hydrateRequestedTaskSessionHistory,
     ensureSessionReadyForView,
-    readSessionFileSearch,
     readSessionModelCatalog,
-    readSessionSlashCommands,
     readSessionTodos,
-    startAgentSession,
-    settleStartedAgentSession,
-    sendAgentMessage,
-    stopAgentSession,
-    updateAgentSessionModel,
-    replyAgentApproval,
-    answerAgentQuestion,
-  } = useAgentOperations();
+  } = agentOperations;
   const sessions = useAgentSessionSummaries();
 
-  const [searchParams, setSearchParams] = useSearchParams();
-  const navigationType = useNavigationType();
-  const [contextSwitchVersion, setContextSwitchVersion] = useState(0);
-  const [selectionIntent, setSelectionIntent] = useState<AgentStudioSelectionIntent | null>(null);
-  const [sessionlessSelection, setSessionlessSelection] =
-    useState<AgentStudioSelectionIntent | null>(null);
   const [gitConflictQuickActionContext, setGitConflictQuickActionContext] =
     useState<AgentStudioGitConflictQuickActionContext | null>(null);
   const gitConflictQuickActionContextRef = useRef<AgentStudioGitConflictQuickActionContext | null>(
     null,
   );
-  const taskDetailsSheetRef = useRef<TaskDetailsSheetControllerHandle | null>(null);
-  const rightPanelRefreshWorktreeRef = useRef<GitDiffRefresh | null>(null);
+  const routeSession = useAgentsPageRouteSessionModel({
+    activeWorkspace,
+    workspaceRepoPath,
+    runtimeDefinitions,
+    isLoadingRuntimeDefinitions,
+    runtimeDefinitionsError,
+    runtimeHealthByRuntime,
+    isLoadingChecks,
+    refreshChecks,
+    refreshRepoRuntimeHealthForRepo,
+    hasCachedRepoRuntimeHealth,
+    tasks,
+    isForegroundLoadingTasks,
+    sessions,
+    hydrateRequestedTaskSessionHistory,
+    ensureSessionReadyForView,
+    readSessionModelCatalog,
+    readSessionTodos,
+  });
   const {
-    taskIdParam,
-    sessionParam,
-    hasExplicitRoleParam,
-    roleFromQuery,
-    isRepoNavigationBoundaryPending,
     navigationPersistenceError,
     retryNavigationPersistence,
-    updateQuery,
-  } = useAgentStudioQuerySync({
-    activeWorkspace,
-    navigationType,
-    searchParams,
-    setSearchParams,
-  });
+    selection,
+    worktreeRecoverySignal,
+  } = routeSession;
 
-  const scheduleQueryUpdate = useCallback(
-    (updates: AgentStudioQueryUpdate): void => {
-      updateQuery(updates);
-    },
-    [updateQuery],
-  );
-
-  const signalContextSwitchIntent = useCallback((): void => {
-    setContextSwitchVersion((current) => current + 1);
-  }, []);
-  const scheduleSelectionIntent = useCallback(
-    (intent: { taskId: string; externalSessionId: string | null; role: AgentRole }): void => {
-      setSelectionIntent(intent);
-      setSessionlessSelection(intent.externalSessionId === null ? intent : null);
-    },
-    [],
-  );
   const handleGitConflictQuickActionContextChange = useCallback(
     (context: AgentStudioGitConflictQuickActionContext | null): void => {
       gitConflictQuickActionContextRef.current = context;
@@ -234,86 +121,6 @@ export function useAgentsPageShellModel(): AgentsPageShellModel {
     },
     [],
   );
-
-  const clearComposerInput = signalContextSwitchIntent;
-  const activeSessionlessSelection =
-    sessionlessSelection &&
-    sessionlessSelection.taskId === taskIdParam &&
-    sessionlessSelection.role === roleFromQuery &&
-    sessionParam === null
-      ? sessionlessSelection
-      : null;
-  const readiness = useAgentStudioReadiness({
-    activeWorkspace,
-    runtimeDefinitions,
-    isLoadingRuntimeDefinitions,
-    runtimeDefinitionsError,
-    runtimeHealthByRuntime,
-    isLoadingChecks,
-    refreshChecks,
-  });
-
-  const selection = useAgentStudioSelectionController({
-    activeWorkspace,
-    isRepoNavigationBoundaryPending,
-    tasks,
-    isLoadingTasks: isForegroundLoadingTasks,
-    sessions,
-    taskIdParam,
-    sessionParam,
-    hasExplicitRoleParam,
-    roleFromQuery,
-    selectionIntent: selectionIntent ?? activeSessionlessSelection,
-    updateQuery: scheduleQueryUpdate,
-    agentStudioReadinessState: readiness.agentStudioReadinessState,
-    hydrateRequestedTaskSessionHistory,
-    ensureSessionReadyForView,
-    runtimeDefinitions,
-    readSessionModelCatalog,
-    readSessionTodos,
-    clearComposerInput,
-    onContextSwitchIntent: signalContextSwitchIntent,
-  });
-
-  useEffect(() => {
-    if (isRepoNavigationBoundaryPending) {
-      setSelectionIntent(null);
-      return;
-    }
-
-    if (!selectionIntent) {
-      return;
-    }
-
-    const selectionIntentResolved = isSelectionIntentResolved({
-      selectionIntent,
-      taskIdParam,
-      sessionParam,
-      roleFromQuery,
-    });
-
-    if (selectionIntentResolved) {
-      setSelectionIntent(null);
-    }
-  }, [isRepoNavigationBoundaryPending, roleFromQuery, selectionIntent, sessionParam, taskIdParam]);
-
-  const isSessionSelectionResolving = Boolean(
-    selectionIntent &&
-      !isRepoNavigationBoundaryPending &&
-      !isSelectionIntentResolved({
-        selectionIntent,
-        taskIdParam,
-        sessionParam,
-        roleFromQuery,
-      }),
-  );
-
-  const openTaskDetails = useCallback((): void => {
-    if (!selection.viewSelectedTask) {
-      return;
-    }
-    taskDetailsSheetRef.current?.openTask(selection.viewSelectedTask.id);
-  }, [selection.viewSelectedTask]);
 
   const handleDetectPullRequest = useCallback(
     (taskId: string): void => {
@@ -337,234 +144,53 @@ export function useAgentsPageShellModel(): AgentsPageShellModel {
     cancelLinkMergedPullRequest();
   }, [cancelLinkMergedPullRequest]);
 
-  const draftStateKey = useMemo(
-    () =>
-      [
-        selection.viewTaskId,
-        selection.viewRole,
-        selection.viewActiveSession?.externalSessionId ?? "new",
-        contextSwitchVersion,
-      ].join(":"),
-    [
-      contextSwitchVersion,
-      selection.viewActiveSession?.externalSessionId,
-      selection.viewRole,
-      selection.viewTaskId,
-    ],
-  );
-
-  const [worktreeRecoverySignal, setWorktreeRecoverySignal] = useState(0);
-  const lastWorktreeRecoveryKeyRef = useRef<string | null>(null);
-
-  useEffect(() => {
-    const nextRecoveryKey = [
-      workspaceRepoPath ?? "",
-      selection.viewTaskId ?? "",
-      selection.viewSelectedTask?.updatedAt ?? "",
-      selection.viewSelectedTask?.status ?? "",
-      selection.viewActiveSession?.externalSessionId ?? "",
-      selection.viewActiveSession?.status ?? "",
-      selection.viewActiveSession?.workingDirectory ?? "",
-      selection.isViewSessionHistoryHydrating ? "1" : "0",
-      isForegroundLoadingTasks ? "1" : "0",
-    ].join(":");
-
-    if (lastWorktreeRecoveryKeyRef.current === null) {
-      lastWorktreeRecoveryKeyRef.current = nextRecoveryKey;
-      return;
-    }
-
-    if (lastWorktreeRecoveryKeyRef.current === nextRecoveryKey) {
-      return;
-    }
-
-    lastWorktreeRecoveryKeyRef.current = nextRecoveryKey;
-    setWorktreeRecoverySignal((previous) => previous + 1);
-  }, [
-    isForegroundLoadingTasks,
-    selection.isViewSessionHistoryHydrating,
-    selection.viewActiveSession?.externalSessionId,
-    selection.viewActiveSession?.status,
-    selection.viewActiveSession?.workingDirectory,
-    selection.viewSelectedTask?.status,
-    selection.viewSelectedTask?.updatedAt,
-    selection.viewTaskId,
-    workspaceRepoPath,
-  ]);
-
-  useEffect(() => {
-    if (!workspaceRepoPath || runtimeDefinitions.length === 0 || isLoadingChecks) {
-      return;
-    }
-    const runtimeKinds = runtimeDefinitions.map((definition) => definition.kind);
-    if (hasCachedRepoRuntimeHealth(workspaceRepoPath, runtimeKinds)) {
-      return;
-    }
-    void refreshRepoRuntimeHealthForRepo(workspaceRepoPath, false);
-  }, [
-    workspaceRepoPath,
-    hasCachedRepoRuntimeHealth,
-    isLoadingChecks,
-    refreshRepoRuntimeHealthForRepo,
-    runtimeDefinitions,
-  ]);
-
-  useAgentStudioQuerySessionSync({
-    isRepoNavigationBoundaryPending,
-    isLoadingTasks: isForegroundLoadingTasks,
+  const { openTaskDetails, mergedPullRequestModal, taskDetailsSheet } = useAgentsPageShellOverlays({
+    activeWorkspace,
     tasks,
-    taskIdParam,
-    sessionParam,
-    selectedSessionById: selection.selectedSessionById,
-    taskId: selection.taskId,
-    activeSession: selection.activeSession,
-    roleFromQuery,
-    isActiveTaskHydrated: selection.isActiveTaskHydrated,
-    scheduleQueryUpdate,
+    selectedTaskId: selection.viewSelectedTask?.id ?? null,
+    pendingMergedPullRequest,
+    linkingMergedPullRequestTaskId,
+    detectingPullRequestTaskId,
+    unlinkingPullRequestTaskId,
+    onDetectPullRequest: handleDetectPullRequest,
+    onUnlinkPullRequest: handleUnlinkPullRequest,
+    onLinkMergedPullRequest: handleLinkMergedPullRequest,
+    onCancelLinkMergedPullRequest: handleCancelLinkMergedPullRequest,
   });
 
-  const orchestration = useAgentStudioOrchestrationController({
+  const {
+    orchestration,
+    orchestrationSelection,
+    handleResolveRebaseConflict,
+    agentStudioHeaderModel,
+  } = useAgentsPageOrchestrationShellModel({
     activeWorkspace,
     branches,
     runtimeDefinitions,
-    selection: {
-      ...selection,
-      contextSwitchVersion,
-      isSessionSelectionResolving,
-      isLoadingTasks: isForegroundLoadingTasks,
-    },
-    readiness,
+    isForegroundLoadingTasks,
+    routeSession,
     hasActiveGitConflict: gitConflictQuickActionContext !== null,
-    draftStateKey,
-    actions: {
-      updateQuery: scheduleQueryUpdate,
-      onContextSwitchIntent: signalContextSwitchIntent,
-      openTaskDetails,
-      startAgentSession,
-      settleStartedAgentSession,
-      sendAgentMessage,
-      stopAgentSession,
-      updateAgentSessionModel,
-      readSessionFileSearch,
-      readSessionSlashCommands,
-      bootstrapTaskSessions,
-      hydrateRequestedTaskSessionHistory,
-      humanRequestChangesTask,
-      setTaskTargetBranch,
-      replyAgentApproval,
-      answerAgentQuestion,
-      scheduleSelectionIntent,
-    },
+    gitConflictQuickActionContext,
+    gitConflictQuickActionContextRef,
+    openTaskDetails,
+    agentOperations,
+    humanRequestChangesTask,
+    setTaskTargetBranch,
   });
 
-  const { startSessionRequest } = orchestration;
-  const activeSessionSummary = useMemo(
-    () =>
-      selection.activeSessionSummary ??
-      (selection.activeSession ? toAgentSessionSummary(selection.activeSession) : null),
-    [selection.activeSession, selection.activeSessionSummary],
-  );
-  const viewActiveSessionSummary = useMemo(
-    () =>
-      selection.viewActiveSessionSummary ??
-      (selection.viewActiveSession ? toAgentSessionSummary(selection.viewActiveSession) : null),
-    [selection.viewActiveSession, selection.viewActiveSessionSummary],
-  );
-  const rebaseConflictSelection = useMemo(
-    () => ({
-      viewTaskId: selection.viewTaskId,
-      viewSelectedTask: selection.viewSelectedTask,
-      viewActiveSession: viewActiveSessionSummary,
-      activeSession: activeSessionSummary,
-      selectedSessionById: selection.selectedSessionById,
-      viewSessionsForTask: selection.viewSessionsForTask,
-      sessionsForTask: selection.sessionsForTask,
-    }),
-    [
-      activeSessionSummary,
-      selection.selectedSessionById,
-      selection.sessionsForTask,
-      selection.viewSelectedTask,
-      selection.viewSessionsForTask,
-      selection.viewTaskId,
-      viewActiveSessionSummary,
-    ],
-  );
-
-  const { handleResolveRebaseConflict } = useAgentStudioRebaseConflictResolution({
+  const { isRightPanelVisible, rightPanelContent } = useAgentsPageRightPanelShellModel({
     activeWorkspace,
-    selection: rebaseConflictSelection,
-    scheduleQueryUpdate,
-    onContextSwitchIntent: signalContextSwitchIntent,
-    startSessionRequest,
+    branches: branches ?? [],
+    activeBranch,
+    selection: orchestrationSelection,
+    orchestration,
+    worktreeRecoverySignal,
+    setTaskTargetBranch,
+    detectingPullRequestTaskId,
+    onDetectPullRequest: handleDetectPullRequest,
+    onResolveGitConflict: handleResolveRebaseConflict,
+    onGitConflictQuickActionContextChange: handleGitConflictQuickActionContextChange,
   });
-
-  const isRightPanelVisible = Boolean(
-    orchestration.rightPanel.panelKind && orchestration.rightPanel.isPanelOpen,
-  );
-  const rightPanelSessionRole = selection.viewActiveSession?.role ?? null;
-  const rightPanelSessionStatus = selection.viewActiveSession?.status ?? null;
-  const rightPanelSessionWorkingDirectory = selection.viewActiveSession?.workingDirectory ?? null;
-  const rightPanelHasActiveSession = selection.viewActiveSession != null;
-  const rightPanelSession = useMemo<BuildToolsSessionDescriptor>(
-    () => ({
-      role: rightPanelSessionRole,
-      status: rightPanelSessionStatus,
-      workingDirectory: rightPanelSessionWorkingDirectory,
-      hasActiveSession: rightPanelHasActiveSession,
-    }),
-    [
-      rightPanelHasActiveSession,
-      rightPanelSessionRole,
-      rightPanelSessionStatus,
-      rightPanelSessionWorkingDirectory,
-    ],
-  );
-  const rightPanelContent = orchestration.rightPanel.panelKind ? (
-    <>
-      <AgentsPageBuildWorktreeRefreshRuntime
-        panelKind={orchestration.rightPanel.panelKind}
-        isPanelOpen={orchestration.rightPanel.isPanelOpen}
-        viewRole={selection.viewRole}
-        activeSession={selection.viewActiveSession}
-        isSessionHistoryHydrating={selection.isViewSessionHistoryHydrating}
-        refreshWorktreeRef={rightPanelRefreshWorktreeRef}
-      />
-      <AgentsPageRightPanelRuntime
-        activeWorkspace={activeWorkspace}
-        branches={branches}
-        activeBranch={activeBranch}
-        viewRole={selection.viewRole}
-        viewTaskId={selection.viewTaskId}
-        session={rightPanelSession}
-        viewSelectedTask={selection.viewSelectedTask}
-        panelKind={orchestration.rightPanel.panelKind}
-        isPanelOpen={orchestration.rightPanel.isPanelOpen}
-        isViewSessionHistoryHydrating={selection.isViewSessionHistoryHydrating}
-        documentsModel={orchestration.agentStudioWorkspaceSidebarModel}
-        repoSettings={orchestration.repoSettings}
-        worktreeRecoverySignal={worktreeRecoverySignal}
-        setTaskTargetBranch={setTaskTargetBranch}
-        detectingPullRequestTaskId={detectingPullRequestTaskId}
-        onDetectPullRequest={handleDetectPullRequest}
-        onResolveGitConflict={handleResolveRebaseConflict}
-        onGitConflictQuickActionContextChange={handleGitConflictQuickActionContextChange}
-        refreshWorktreeRef={rightPanelRefreshWorktreeRef}
-      />
-    </>
-  ) : null;
-
-  const agentStudioHeaderModel = useMemo(
-    () => ({
-      ...orchestration.agentStudioHeaderModel,
-      onResolveGitConflictQuickAction: gitConflictQuickActionContext
-        ? () => {
-            void gitConflictQuickActionContextRef.current?.resolveWithBuilder();
-          }
-        : null,
-    }),
-    [gitConflictQuickActionContext, orchestration.agentStudioHeaderModel],
-  );
 
   const sessionStartModal = orchestration.sessionStartModal ? (
     <SessionStartModal model={orchestration.sessionStartModal} />
@@ -572,31 +198,6 @@ export function useAgentsPageShellModel(): AgentsPageShellModel {
 
   const humanReviewFeedbackModal = (
     <HumanReviewFeedbackModal model={orchestration.humanReviewFeedbackModal} />
-  );
-
-  const mergedPullRequestModal = pendingMergedPullRequest ? (
-    <MergedPullRequestConfirmDialog
-      pullRequest={pendingMergedPullRequest.pullRequest}
-      isLinking={pendingMergedPullRequest.taskId === linkingMergedPullRequestTaskId}
-      onCancel={handleCancelLinkMergedPullRequest}
-      onConfirm={() => void handleLinkMergedPullRequest()}
-    />
-  ) : null;
-
-  const taskDetailsSheet = (
-    <TaskDetailsSheetController
-      ref={taskDetailsSheetRef}
-      activeWorkspace={activeWorkspace}
-      allTasks={tasks}
-      taskSessionsByTaskId={EMPTY_TASK_SESSIONS_BY_TASK_ID}
-      activeTaskSessionContextByTaskId={EMPTY_ACTIVE_TASK_SESSION_CONTEXT_BY_TASK_ID}
-      workflowActionsEnabled={false}
-      onOpenSession={noopOpenSession}
-      onDetectPullRequest={handleDetectPullRequest}
-      onUnlinkPullRequest={handleUnlinkPullRequest}
-      detectingPullRequestTaskId={detectingPullRequestTaskId}
-      unlinkingPullRequestTaskId={unlinkingPullRequestTaskId}
-    />
   );
 
   return {
